% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_methods.R
\name{application_methods}
\alias{application_methods}
\title{Aplica os métodos de projeção}
\usage{
application_methods(census, anos, assin)
}
\arguments{
\item{census}{banco de dados de população para 3 censos}

\item{anos}{vetor com os anos do primeiro e segundo censo e do ano a ser projetado}
}
\value{
População projetada n anos
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{
data(brasil2000)
}
