% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_erros.R
\name{calc_erros}
\alias{calc_erros}
\title{Calcula estrutura de erros}
\usage{
calc_erros(dados)
}
\arguments{
\item{dados}{banco de dados de resultados}
}
\value{
Lista com informações sobre erros
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{
data(brasil2000)
}
