% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_aibi.R
\name{forecast_aibi}
\alias{forecast_aibi}
\title{Função para projeção linear (Share of Growth) - AiBi}
\usage{
forecast_aibi(census1, census2, UF_census1, UF_census2, UF_proj_2022)
}
\arguments{
\item{census1}{população do primeiro censo}

\item{census2}{população do segundo censo}

\item{anos}{vetor com os anos do primeiro e segundo censo e do ano a ser projetado}
}
\value{
População projetada n anos
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{
data(brasil2000)
}
