% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_linear.R
\name{forecast_linear}
\alias{forecast_linear}
\title{Função para projeção linear}
\usage{
forecast_linear(census1, census2, anos)
}
\arguments{
\item{census1}{população do primeiro censo}

\item{census2}{população do segundo censo}

\item{anos}{vetor com os anos do primeiro e segundo censo e do ano a ser projetado}
}
\value{
População projetada n anos
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{
data(brasil2000)
}
