% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_linear_prop.R
\name{forecast_linear_prop}
\alias{forecast_linear_prop}
\title{Função para projeção linear de proporções}
\usage{
forecast_linear_prop(prop_census1, prop_census2, anos)
}
\arguments{
\item{anos}{vetor com os anos do primeiro e segundo censo e do ano a ser projetado}

\item{census1}{proporção de população do primeiro censo}

\item{census2}{proporção de população do segundo censo}
}
\value{
População projetada n anos
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{
data(brasil2000)
}
