% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_estimativas.R
\name{forecast_logistico}
\alias{forecast_logistico}
\title{Função para projeção logística}
\usage{
forecast_logistico(census1, census2, anos, assint)
}
\arguments{
\item{census1}{The temperature in degrees Fahrenheit}
}
\value{
População projetada n anos
}
\description{
Gera projeção de população total logística
}
\examples{
data(brasil2000)
}
