% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{n_projection}
\alias{n_projection}
\title{Aplica os métodos de projeção}
\usage{
n_projection(census, anos, DAM_proj, ano_proj, assint)
}
\arguments{
\item{census}{banco de dados de população para 3 censos}

\item{anos}{vetor com os anos do primeiro e segundo censo e do ano a ser projetado}
}
\value{
População projetada n anos
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{
data(brasil2000)
}
