% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_rates.R
\name{plot_growth_rates}
\alias{plot_growth_rates}
\title{Plota}
\usage{
plot_growth_rates(
  censos,
  census_date,
  pais_name = NULL,
  DAM_name = NULL,
  DAMe_name = NULL,
  save_plotly = T,
  save_image = T,
  path = getwd()
)
}
\arguments{
\item{dados}{banco de dados de resultados}
}
\value{
Lista com informações sobre erros
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{
data(brasil2000)
}
