% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_estimativas.R, R/project_exp.R
\name{project_exp}
\alias{project_exp}
\title{Função para projeção exponencial usando a fórmula fornecida e anos flexíveis}
\usage{
project_exp(census1, census2, anos)

project_exp(census1, census2, anos)
}
\arguments{
\item{census1}{população do primeiro censo}

\item{census2}{população do segundo censo}

\item{anos}{vetor com os anos do primeiro e segundo censo e do ano a ser projetado}
}
\value{
População projetada n anos

População projetada n anos
}
\description{
Gera rojeção exponencial usando a fórmula fornecida e anos flexíveis

Gera rojeção exponencial usando a fórmula fornecida e anos flexíveis
}
\examples{
data(brasil2000)
data(brasil2000)
}
