% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_logistico.R, R/funcoes_estimativas.R
\name{forecast_logistico}
\alias{forecast_logistico}
\title{Função para projeção logística}
\usage{
forecast_logistico(census1, census2, anos, assint)

forecast_logistico(census1, census2, anos, assint)
}
\arguments{
\item{census1}{população do primeiro censo}

\item{census2}{população do segundo censo}

\item{anos}{vetor com os anos do primeiro e segundo censo e do no a ser projetado}

\item{assint}{limite inferior e superior das assintotas da curva logística}
}
\value{
População projetada n anos

População projetada n anos
}
\description{
Gera projeção de população total logística

Gera projeção de população total logística
}
\examples{
data(brasil2000)
data(brasil2000)
}
