% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_estimativas.R
\name{forecast_logistico}
\alias{forecast_logistico}
\title{Função para projeção logística}
\usage{
forecast_logistico(census1, census2, anos, assint)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Gera prpjção de população total logística
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
